+++ janitor/functions.py
@@ -2980,20 +2980,14 @@ def dropnotnull(df: pd.DataFrame, column_name: Hashable) -> pd.DataFrame:
 
 
 @pf.register_dataframe_method
-@deprecated_alias(column="column_name")
-def find_replace(
-    df: pd.DataFrame, column_name: str, mapper: Dict, match: str = "exact"
-) -> pd.DataFrame:
+def find_replace(df, match: str = "exact", **mappings):
     """
-    Perform a find-and-replace action on a column of data.
+    Perform a find-and-replace action on provided columns.
 
-    This method mutates the original DataFrame.
-
-    Depending on use cases, users can choose either exact, full value matching
-    or regular-expression-based fuzzy matching (substring matching is allowed
-    in the latter case). For strings, the matching is always case sensitive.
-
-    Note that default value for keyword argument `match` is "exact".
+    Depending on use case, users can choose either exact, full-value matching,
+    or regular-expression-based fuzzy matching
+    (hence allowing substring matching in the latter case).
+    For strings, the matching is always case sensitive.
 
     For instance, given a dataframe containing orders at a coffee shop:
 
@@ -3013,27 +3007,77 @@ def find_replace(
 
         # Functional usage
         df = find_replace(
-            df, 'order', {'ice coffee': 'latte', 'regular coffee': 'latte'},
-            match='exact'
+            df,
+            match='exact',
+            order={'ice coffee': 'latte', 'regular coffee': 'latte'},
         )
 
         # Method chaining usage
         df = df.find_replace(
-            'order', {'ice coffee': 'latte', 'regular coffee': 'latte'},
             match='exact'
+            order={'ice coffee': 'latte', 'regular coffee': 'latte'},
         )
-
 
     Example 2: Regular-expression-based matching
 
     .. code-block:: python
 
         # Functional usage
-        df = find_replace(df, 'order', {'coffee$': 'latte'}, match='regex')
+        df = find_replace(
+            df,
+            match='regex',
+            order={'coffee$': 'latte'},
+        )
 
         # Method chaining usage
-        df = df.find_replace('order', {'coffee$': 'latte'}, match='regex')
+        df = df.find_replace(
+            match='regex',
+            order={'coffee$': 'latte'},
+        )
 
+    To perform a find and replace on the entire dataframe,
+    pandas' ``df.replace()`` function provides the appropriate functionality.
+    You can find more detail on the replace_ docs.
+
+    This function only works with column names that have no spaces
+    or punctuation in them.
+    For example, a column name ``item_name`` would work with ``find_replace``,
+    because it is a contiguous string that can be parsed correctly,
+    but ``item name`` would not be parsed correctly by the Python interpreter.
+
+    If you have column names that might not be compatible,
+    we recommend calling on ``clean_names()`` as the first method call.
+    If, for whatever reason, that is not possible,
+    then ``_find_replace()`` is available as a function
+    that you can do a pandas pipe_ call on.
+
+    .. _replace: https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.replace.html
+    .. _pipe: https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.pipe.html
+
+    :param df: A pandas DataFrame.
+    :param match: Whether or not to perform an exact match or not.
+        Valid values are "exact" or "regex".
+    :param mappings: keyword arguments corresponding to column names
+        that have dictionaries passed in indicating what to find (keys)
+        and what to replace with (values).
+    """  # noqa: E501
+    for column_name, mapper in mappings.items():
+        df = _find_replace(df, column_name, mapper, match=match)
+    return df
+
+
+def _find_replace(
+    df: pd.DataFrame, column_name: str, mapper: Dict, match: str = "exact"
+) -> pd.DataFrame:
+    """
+    Utility function for ``find_replace``.
+
+    The code in here was the original implementation of ``find_replace``,
+    but we decided to change out the front-facing API to accept
+    kwargs + dictionaries for readability,
+    and instead dispatch underneath to this function.
+    This implementation was kept
+    because it has a number of validations that are quite useful.
 
     :param df: A pandas DataFrame.
     :param column_name: The column on which the find/replace action is to be